+++ site/configs/theme.ts
@@ -1,6 +1,9 @@
 // Types
 import type { Theme } from '@/types/configs/theme'
 
-const themeConfig: Theme = { color: '#e39d52' }
+const themeConfig: Theme = {
+	color: '#b96',
+	drawerWidth: 240
+}
 
 export default themeConfig