+++ site/types/components/atoms/mode-button.d.ts
@@ -1,0 +1,7 @@
+export type ModeButton = {
+	mode: boolean
+	label?: string
+	ariaLabel?: string
+	icon?: JSX.Element
+	onClick?: () => void
+}