+++ site/types/components/molecules/drawer.d.ts
@@ -1,0 +1,7 @@
+// Types
+import type { Dispatch, SetStateAction } from 'react'
+
+export type DrawerProps = {
+	open: boolean
+	setOpen: Dispatch<SetStateAction>
+}