+++ site/types/components/molecules/mode.d.ts
@@ -1,0 +1,7 @@
+// Types
+import type { SxProps } from '@mui/material'
+
+export type ModeProps = {
+	text: boolean
+	sx?: SxProps
+}