+++ CHANGES.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 - Changed default colors
+- Created `header` and `drawer`
 
 ## 0.0.1
 - Initial commit